requirements.txt
alpaca-py>=0.21.0
openai>=1.30.0
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
